# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/json-dto/endpoints": {
      "get": {
        "operationId": "JsonDtoController_getEndpoints",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of endpoints with their details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JsonDtoEndpointDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all JSON DTO endpoints with responses",
        "tags": [
          "Json DTO Inspector"
        ]
      }
    },
    "/json-dto/list": {
      "get": {
        "operationId": "JsonDtoController_listDtos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of DTO names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all available DTOs",
        "tags": [
          "Json DTO Inspector"
        ]
      }
    },
    "/json-dto/schema": {
      "get": {
        "operationId": "JsonDtoController_getSchema",
        "parameters": [
          {
            "name": "dto",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON schema of the selected DTO",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonDtoSchemaResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get JSON schema of a selected DTO",
        "tags": [
          "Json DTO Inspector"
        ]
      }
    },
...
    "/api/schema": {
      "post": {
        "operationId": "SchemaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchemaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSchemaDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed."
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient role"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Create a new Schema",
        "tags": [
          "Schema"
        ]
      },
      "get": {
        "operationId": "SchemaController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of Schema records.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CreateSchemaDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - insufficient role"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "bearer": []
          }
        ],
        "summary": "Retrieve all Schema records",
        "tags": [
          "Schema"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Server",
    "description": "Authentication and Role Protected APIs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token in the format: Bearer <token>",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt",
        "description": "JWT stored in cookie (for web clients)"
      }
    },
    "schemas": {
...
      "JsonDtoSchemaResponseDto": {
        "type": "object",
        "properties": {
          "schema": {
            "type": "object",
            "description": "Generated JSON schema of the specified DTO",
            "additionalProperties": {
              "type": "object"
            }
          }
        },
        "required": [
          "schema"
        ]
      },
      "JsonDtoEndpointQueryParamDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the query parameter",
            "example": "dto"
          },
          "required": {
            "type": "boolean",
            "description": "Whether this query parameter is required",
            "example": false
          },
          "description": {
            "type": "string",
            "description": "Description of the query parameter",
            "example": "Type of DTO to fetch the schema for"
          }
        },
        "required": [
          "name",
          "required",
          "description"
        ]
      },
      "JsonDtoEndpointResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "description": "HTTP status code",
            "example": 200
          },
          "description": {
            "type": "string",
            "description": "Description of the response",
            "example": "Successful response with DTO schema"
          },
          "schema": {
            "type": "object",
            "description": "Schema or type of the response payload",
            "additionalProperties": true,
            "example": {
              "schema": {
                "property1": {
                  "type": "string"
                },
                "property2": {
                  "type": "number"
                }
              }
            }
          }
        },
        "required": [
          "status",
          "description",
          "schema"
        ]
      },
      "JsonDtoEndpointDto": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string",
            "description": "HTTP method for the endpoint",
            "example": "GET"
          },
          "path": {
            "type": "string",
            "description": "Full URL path of the endpoint",
            "example": "/json-dto/custom"
          },
          "queryParams": {
            "description": "List of query parameters accepted by the endpoint",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JsonDtoEndpointQueryParamDto"
            }
          },
          "description": {
            "type": "string",
            "description": "Short description of the endpoint",
            "example": "Returns the JSON schema for a selected DTO"
          },
          "responses": {
            "description": "List of possible responses for the endpoint",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JsonDtoEndpointResponseDto"
            }
          }
        },
        "required": [
          "method",
          "path",
          "queryParams",
          "description",
          "responses"
        ]
      },
...
      "CreateSchemaDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "name field"
          },
          "schema": {
            "type": "object",
            "description": "schema field"
          }
        },
        "required": [
          "name",
          "schema"
        ]
      },
...
    }
  }
}